+++ generators/gen_models.py
@@ -17,7 +17,10 @@ def main():
         input_=OPEN_API_PATH,
         input_file_type=dcg.InputFileType.OpenAPI,
         use_double_quotes=True,
+        use_standard_collections=True,
+        use_schema_description=True,
         set_default_enum_member=True,
+        target_python_version=dcg.PythonVersion.PY_310,
         output_model_type=dcg.DataModelType.PydanticV2BaseModel,
         output=MODELS_PATH,
     )