+++ s2gos/client/gui/jobstable.py
@@ -10,11 +10,11 @@ import param
 
 from s2gos.client import ClientException
 from s2gos.common.models import (
+    InlineOrRefValue,
+    JobInfo,
     JobList,
+    ProcessResults,
     StatusCode,
-    StatusInfo,
-    Results,
-    InlineOrRefData,
 )
 
 JobAction: TypeAlias = Callable[[str], Any]
@@ -127,11 +127,11 @@ class JobsTable(pn.viewable.Viewer):
         )
 
     @classmethod
-    def is_disabled(cls, jobs: list[StatusInfo], requirements: set[StatusCode]):
+    def is_disabled(cls, jobs: list[JobInfo], requirements: set[StatusCode]):
         return not jobs or not all(j.status in requirements for j in jobs)
 
     @property
-    def selected_jobs(self) -> list[StatusInfo]:
+    def selected_jobs(self) -> list[JobInfo]:
         """Get selected jobs from jobs table."""
         selection = self._tabulator.selection
         if not selection:
@@ -161,15 +161,15 @@ class JobsTable(pn.viewable.Viewer):
         )
 
     def _on_get_job_result_clicked(self, _event: Any):
-        def handle_result(_job_id: str, results: Results | dict):
+        def handle_result(_job_id: str, results: ProcessResults | dict):
             # noinspection PyProtectedMember
             from IPython import get_ipython
 
-            if isinstance(results, Results):
+            if isinstance(results, ProcessResults):
                 results = results.root
             if isinstance(results, dict):
                 results = {
-                    k: (v.root if isinstance(v, InlineOrRefData) else v)
+                    k: (v.root if isinstance(v, InlineOrRefValue) else v)
                     for k, v in results.items()
                 }
             var_name = "_results"
@@ -257,11 +257,11 @@ class JobsTable(pn.viewable.Viewer):
         return tabulator
 
     @classmethod
-    def _jobs_to_dataframe(cls, jobs: list[StatusInfo]):
+    def _jobs_to_dataframe(cls, jobs: list[JobInfo]):
         return pd.DataFrame([cls._job_to_dataframe_row(job) for job in jobs])
 
     @classmethod
-    def _job_to_dataframe_row(cls, job: StatusInfo):
+    def _job_to_dataframe_row(cls, job: JobInfo):
         return {
             "process_id": job.processID,
             "job_id": job.jobID,