+++ s2gos/server/services/local/job.py
@@ -4,14 +4,15 @@
 
 import datetime
 import inspect
+import traceback
 import warnings
 from abc import ABC, abstractmethod
 from concurrent.futures import Future
 from typing import Any, Callable, Optional
 
 from s2gos.common.models import (
+    JobInfo,
     StatusCode,
-    StatusInfo,
     Type,
 )
 
@@ -74,7 +75,7 @@ class Job(JobContext):
         function: Callable[..., Any],
         function_kwargs: dict[str, Any],
     ):
-        self.status_info = StatusInfo(
+        self.status_info = JobInfo(
             type=Type.process,
             processID=process_id,
             jobID=job_id,
@@ -140,8 +141,10 @@ class Job(JobContext):
         self.status_info.finished = datetime.datetime.now()
         self.status_info.status = status_code
         if exception is not None:
-            # TODO: save exception traceback
             self.status_info.message = f"{exception}"
+            self.status_info.traceback = traceback.format_exception(
+                type(exception), exception, exception.__traceback__
+            )
 
 
 def get_job_context() -> JobContext: