+++ tests/client/test_client.py
@@ -7,14 +7,13 @@ from unittest import TestCase
 from s2gos.client import Client
 from s2gos.common.models import (
     ConfClasses,
-    Execute,
-    InlineOrRefData,
+    JobInfo,
     JobList,
     LandingPage,
-    Process,
+    ProcessDescription,
     ProcessList,
-    Results,
-    StatusInfo,
+    ProcessRequest,
+    ProcessResults,
 )
 from tests.client.helpers import MockTransport
 
@@ -48,16 +47,17 @@ class ClientTest(TestCase):
 
     def test_get_process_description(self):
         result = self.client.get_process_description(process_id="gobabeb_1")
-        self.assertIsInstance(result, Process)
+        self.assertIsInstance(result, ProcessDescription)
 
     def test_execute(self):
         result = self.client.execute(
             process_id="gobabeb_1",
-            request=Execute(
-                inputs={"bbox": InlineOrRefData(root=[10, 20, 30, 40])}, outputs={}
+            request=ProcessRequest(
+                inputs={"bbox": [10, 20, 30, 40]},
+                outputs={},
             ),
         )
-        self.assertIsInstance(result, StatusInfo)
+        self.assertIsInstance(result, JobInfo)
 
     def test_get_jobs(self):
         result = self.client.get_jobs()
@@ -65,12 +65,12 @@ class ClientTest(TestCase):
 
     def test_dismiss(self):
         result = self.client.dismiss("job_12")
-        self.assertIsInstance(result, StatusInfo)
+        self.assertIsInstance(result, JobInfo)
 
     def test_get_status(self):
         result = self.client.get_status("job_12")
-        self.assertIsInstance(result, StatusInfo)
+        self.assertIsInstance(result, JobInfo)
 
     def test_get_result(self):
         result = self.client.get_result("job_12")
-        self.assertIsInstance(result, Results)
+        self.assertIsInstance(result, ProcessResults)